# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  models.BalanceOperation:
    properties:
      account_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      new_balance:
        type: number
      operation_id:
        type: string
      operation_type:
        type: string
      payment_id:
        type: string
      processed_at:
        type: string
      status:
        type: string
    type: object
...
  models.ConvertCurrencyResponse:
    properties:
      operation_id:
        type: string
    type: object
...
  models.CurrencyRateResponse:
    properties:
      from_currency:
        type: string
      rate:
        type: number
      to_currency:
        type: string
    type: object
...
  models.DepositResponse:
    properties:
      payment_id:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      receiver:
        type: string
      sender:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateCurrencyRateRequest:
    properties:
      from_currency:
        type: string
      rate:
        type: number
      to_currency:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: API для платежной системы
  title: Payment API
  version: "1.0"
paths:
...
  /payments/{paymentID}:
    get:
      description: Возвращает информацию о конкретном платеже
      parameters:
      - description: ID платежа
        in: path
        name: paymentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить платеж
      tags:
      - payments
  /payments/{paymentID}/cancel:
    patch:
      description: Отменяет указанный платеж
      parameters:
      - description: ID платежа
        in: path
        name: paymentID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Отменить платеж
      tags:
      - payments
  /payments/{paymentID}/update:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Обновляет статус указанного платежа
      parameters:
      - description: ID платежа
        in: path
        name: paymentID
        required: true
        type: string
      - description: Статус
        enum:
        - completed
        - refunded
        in: formData
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Обновить статус платежа
      tags:
      - payments
  /payments/convert:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Конвертирует указанную сумму из одной валюты в другую
      parameters:
      - description: Номер счёта с валютой для обмена
        in: formData
        name: sender
        required: true
        type: string
      - description: Номер счёта для зачисления новой валюты
        in: formData
        name: receiver
        required: true
        type: string
      - description: Сумма платежа
        in: formData
        name: amount
        required: true
        type: number
      - description: Описание платежа
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConvertCurrencyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Конвертировать валюту
      tags:
      - payments
  /payments/currency-rate:
    post:
      consumes:
      - application/json
      description: Обновляет курс конвертации между валютами (Admin only)
      parameters:
      - description: Данные для обновления курса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCurrencyRateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Обновить курс валют
      tags:
      - payments
  /payments/currency-rate/get:
    get:
      description: Возвращает текущий курс обмена между указанными валютами
      parameters:
      - description: 'Исходная валюта (например: USD)'
        in: query
        name: from
        required: true
        type: string
      - description: 'Целевая валюта (например: EUR)'
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CurrencyRateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить курс валют
      tags:
      - payments
  /payments/deposit:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Пополняет счет на указанную сумму
      parameters:
      - description: Номер счета
        in: formData
        name: account_id
        required: true
        type: string
      - description: Сумма платежа
        in: formData
        name: amount
        required: true
        type: number
      - description: Описание платежа
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DepositResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Пополнить счет
      tags:
      - payments
  /payments/get:
    get:
      description: Возвращает список всех платежей
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить все платежи
      tags:
      - payments
  /payments/history:
    get:
      description: Возвращает историю операций по указанному счету
      parameters:
      - description: Номер счета
        in: query
        name: account_id
        required: true
        type: string
      - description: Лимит (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BalanceOperation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить историю операций
      tags:
      - payments
  /payments/transfer:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Создаеёт перевод между указанными счетами
      parameters:
      - description: Платёжный счет
        in: formData
        name: sender
        required: true
        type: string
      - description: Номер счета получателя
        in: formData
        name: receiver
        required: true
        type: string
      - description: Сумма платежа
        in: formData
        name: amount
        required: true
        type: number
      - description: Валюта
        enum:
        - USD
        - EUR
        - RUB
        in: formData
        name: currency
        required: true
        type: string
      - description: Описание платежа
        in: formData
        name: description
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Перевести на другой счет
      tags:
      - payments
...
schemes:
- https
securityDefinitions:
  CookieAuth:
    description: Аутентификация через куки (токен автоматически отправляется сервером)
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"